Added the edit post feature.
We added this first since we want to redirect user to EditPost once they
published their post. We'll finalizing the post creation process after
this.

We also change the newpost.css and js to post-editor, since it's used in
both new and edit post. Edit post will have an additional field of post
ID for submitting though, which we'll need to add on later.